Solves #6 - Ensure backend definition is set
If the backend is unspecified, imgui will do a simple linear check to see which glewalike to use, starting with glew. This is fine in a simple container with the minimum required installs, but if your system has glew it will make the wrong selection.

So this change introduces a CMakeLists.txt parameter to make users aware that there is a choice to be had, but then hardcodes it to a simple default.