Set include dirs to be SYSTEM to prevent user-enabled warnings in header
Currently, the header directories are included normally, which can allow warnings to show up depending on the user's current settings.
Changes the public includes to be system so no warnings are triggered.
Adds a copy of the includes privately so building the library itself still shows the proper warnings.
Also changes glfw's include path to only be added if the variable exists, preventing user's project from including from /include.

Signed-off-by: Gabe Gonzalez II <lilggamegenius@gmail.com>